media_downloader/requirements.txt
altair==4.2.0
anyio==3.6.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.1.0
async-timeout==4.0.2
attrs==22.1.0
autopep8==2.0.0
av==9.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
biopandas==0.4.1
biopython==1.79
bleach==5.0.1
blinker==1.5
Brotli==1.0.9
bs4==0.0.1
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
contourpy==1.0.6
cssselect==1.2.0
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.6
entrypoints==0.4
executing==1.2.0
fake-useragent==1.1.1
fastapi==0.78.0
fastjsonschema==2.16.2
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.8.0
Flask==2.2.2
fonttools==4.38.0
future==0.18.2
get-video-properties==0.1.1
gitdb==4.0.10
GitPython==3.1.29
greenlet==1.1.3
h11==0.13.0
html5lib==1.1
idna==3.4
imageio==2.22.4
imageio-ffmpeg==0.4.7
importlib-metadata==5.1.0
ipykernel==6.17.0
ipyspeck==0.6.1
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
jupyter-server==1.21.0
jupyter_client==7.4.4
jupyter_core==4.11.2
jupyter_packaging==0.12.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
lxml==4.9.2
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
MouseInfo==0.1.3
mpmath==1.2.1
mutagen==1.46.0
names==0.3.0
nbclassic==0.4.7
nbclient==0.7.0
nbconvert==7.2.3
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.0
numpy==1.23.4
OmegaFold==0.0.0
packaging==21.3
pafy==0.5.5
pandas==1.5.1
pandocfilters==1.5.0
parse==1.19.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.0
prometheus-client==0.15.0
prompt-toolkit==3.0.31
protobuf==3.20.3
psutil==5.9.3
pure-eval==0.2.2
py3Dmol==1.8.1
pyarrow==10.0.1
PyAutoGUI==0.9.53
pycodestyle==2.10.0
pycparser==2.21
pycryptodomex==3.16.0
pydantic==1.10.2
pydeck==0.8.0
pydub==0.25.1
pyee==8.2.2
pygame==2.1.2
PyGetWindow==0.0.9
Pygments==2.13.0
pymolPy3==0.1.2
Pympler==1.0.1
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
pyppeteer==1.0.2
pyquery==1.4.3
PyRect==0.2.0
pyrsistent==0.19.1
PyScreeze==0.1.28
python-dateutil==2.8.2
python-dotenv==0.21.0
python-slugify==6.1.2
pytube==12.1.0
pytweening==1.0.4
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pywhatkit==5.4
pywin32==304
pywinpty==2.0.9
pyzmq==24.0.1
rapidfuzz==2.13.6
rdkit==2022.9.1
redis==4.4.0
requests==2.28.1
requests-html==0.10.0
rich==12.6.0
scikit-learn==1.2.0
scipy==1.9.3
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0.post1
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
spotdl==4.0.6
spotipy==2.22.0
SQLAlchemy==1.4.41
stack-data==0.6.0
starlette==0.19.1
stmol==0.0.9
streamlit==1.15.2
sympy==1.11.1
terminado==0.17.0
text-unidecode==1.3
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
toolz==0.12.0
torch==1.13.0+cu117
torchaudio==0.13.0+cu117
torchvision==0.14.0+cu117
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
typing_extensions==4.3.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.12
uvicorn==0.17.6
validators==0.20.0
virtualenv==20.16.6
w3lib==2.1.1
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
websockets==10.4
Werkzeug==2.2.2
widgetsnbextension==3.5.2
wikipedia==1.4.0
youtube-dl==2021.12.17
yt-dlp==2022.11.11
ytmusicapi==0.22.0
zipp==3.11.0
